Build time bridge between Navigator and Player.
