+++ src/components/Navigator/NavigatorCue.tsx
@@ -1,126 +1,35 @@
+import { useViewerDispatch } from "context/viewer-context";
 import React from "react";
-import { styled, keyframes } from "@stitches/react";
 import { cleanTime } from "services/utils";
-import { theme } from "theme";
+import { NavigatorCueAnchor } from "./NavigatorCue.styled";
 
 interface Props {
   label: string;
   startTime: string;
-  t: number;
+  active: boolean;
+  time: number;
 }
 
-const NavigatorCue: React.FC<Props> = ({ label, startTime }) => {
+const NavigatorCue: React.FC<Props> = ({ label, startTime, active, time }) => {
+  const dispatch: any = useViewerDispatch();
+
+  const handleClick = () => {
+    dispatch({
+      type: "updateTime",
+      time: time,
+    });
+  };
+
   return (
-    <NavigatorCueAnchor data-testid="navigator-cue">
+    <NavigatorCueAnchor
+      onClick={handleClick}
+      aria-selected={active}
+      data-testid="navigator-cue"
+    >
       {label}
       <strong>{cleanTime(startTime)}</strong>
     </NavigatorCueAnchor>
   );
 };
-
-const spin = keyframes({
-  from: { transform: "rotate(360deg)" },
-  to: { transform: "rotate(0deg)" },
-});
-
-const NavigatorCueAnchor = styled("a", {
-  display: "flex",
-  flexGrow: "1",
-  justifyContent: "space-between",
-  margin: "0",
-  padding: "0.55rem 1.618rem 0.45rem",
-  lineHeight: "1.45em",
-  fontSize: "1rem",
-  cursor: "pointer",
-  color: theme.color.primaryMuted,
-  position: "relative",
-
-  "&::before": {
-    content: "",
-    width: "12px",
-    height: "12px",
-    borderRadius: "12px",
-    position: "absolute",
-    backgroundColor: theme.color.primaryMuted,
-    opacity: "0",
-    left: "8px",
-    marginTop: "5px",
-  },
-
-  "&::after": {
-    content: "",
-    width: "4px",
-    height: "6px",
-    position: "absolute",
-    backgroundColor: theme.color.secondary,
-    opacity: "0",
-    clipPath: "polygon(100% 50%, 0 100%, 0 0)",
-    left: "13px",
-    marginTop: "8px",
-  },
-
-  strong: {
-    marginLeft: "1rem",
-  },
-
-  "&:hover": {
-    color: theme.color.accent,
-
-    "&::before": {
-      backgroundColor: theme.color.accent,
-      opacity: "1",
-    },
-
-    "&::after": {
-      content: "",
-      width: "4px",
-      height: "6px",
-      position: "absolute",
-      backgroundColor: theme.color.secondary,
-      clipPath: "polygon(100% 50%, 0 100%, 0 0)",
-      opacity: "1",
-    },
-  },
-
-  "&:first-child": {
-    color: theme.color.primary,
-    backgroundColor: theme.color.secondaryMuted,
-
-    "&::before": {
-      content: "",
-      width: "6px",
-      height: "6px",
-      position: "absolute",
-      backgroundColor: "transparent",
-      border: `3px solid ${theme.color.accentMuted}`,
-      borderRadius: "12px",
-      left: "8px",
-      marginTop: "4px",
-      opacity: "1",
-      animation: "1s linear infinite",
-      animationName: spin,
-    },
-
-    "&::after": {
-      content: "",
-      width: "6px",
-      height: "6px",
-      position: "absolute",
-      backgroundColor: "transparent",
-      border: `3px solid ${theme.color.accent}`,
-      clipPath: "polygon(100% 0, 100% 100%, 0 0)",
-      borderRadius: "12px",
-      left: "8px",
-      marginTop: "4px",
-      opacity: "1",
-      animation: "1.5s linear infinite",
-      animationName: spin,
-    },
-  },
-
-  "&:last-child": {
-    margin: "0 0 1.618rem",
-  },
-});
 
 export default NavigatorCue;
