+++ src/dev.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import App from "./index";
 
 const sampleManifest: string =
-  "http://127.0.0.1:8080/fixtures/iiif/manifests/updatedSupplementing.json";
+  "http://127.0.0.1:8080/fixtures/iiif/manifests/assortedCanvases.json";
 
 ReactDOM.render(
   <App manifestId={sampleManifest} />,
